[ao][fx]refactor memoryless from user facing APIs
removed memoryless from user facing APIs and replaced with logic through the use of compute_dtype.
if compute_dtype isn't specified there are no changes to how this works, if compute_dtype is specified
then its used as the observer dtype (rather than the dtype which is treated as the activation_dtype).
Finally if compute_dtype is a quantized dtype and dtype=float (which is the designation for dynamic quantization we
came up with) then a memoryless observer is created.

ghstack-source-id: 58abaca178c4e3808b3735d08a46f33d7245101e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73560